requirements.txt
# ===============================
# KRISHIKAVACH AI - LATEST VERSIONS
# ===============================

# 🔧 CORE WEB FRAMEWORK - LATEST
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3

# 🧠 MACHINE LEARNING & DATA SCIENCE - LATEST
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
scipy==1.11.4

# 🤖 GOOGLE AI & LANGUAGE MODELS - LATEST
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0
google-cloud-aiplatform==1.38.1
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# 🖼️ IMAGE PROCESSING & COMPUTER VISION - LATEST
Pillow==10.1.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
imageio==2.33.0
scikit-image==0.22.0

# 🎙️ AUDIO PROCESSING & SPEECH - LATEST
SpeechRecognition==3.10.0
pydub==0.25.1
pyaudio==0.2.11
webrtcvad==2.0.10

# 📊 DATA VISUALIZATION - LATEST
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
bokeh==3.3.2

# 🗄️ DATABASE & ORM (Optional - LATEST)
# Flask-SQLAlchemy==3.1.1
# SQLAlchemy==2.0.23
# Flask-Migrate==4.0.5
# Flask-Login==0.6.3

# 📧 EMAIL & COMMUNICATION (Optional - LATEST)
# Flask-Mail==0.9.1
# Flask-WTF==1.2.1
# WTForms==3.1.1

# 🔐 SECURITY & AUTH (Optional - LATEST)
# bcrypt==4.0.1
# PyJWT==2.8.0
# cryptography==41.0.7

# 🌍 ENVIRONMENT & CONFIG - LATEST
python-dotenv==1.0.0
python-decouple==3.8

# 🌐 HTTP CLIENT & API - LATEST
requests==2.31.0
urllib3==2.1.0
httpx==0.25.2

# 📁 FILE HANDLING & UTILITIES - LATEST
base64io==1.0.3
pathlib2==2.3.7
path.py==12.5.0

# 📝 LOGGING & MONITORING - LATEST
loguru==0.7.2

# 🧪 TESTING & DEVELOPMENT - LATEST
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
coverage==7.3.2

# 🔄 ASYNC & CONCURRENCY (Optional - LATEST)
# asyncio==3.4.3
# aiohttp==3.9.1
# celery==5.3.4

# 📱 MOBILE & RESPONSIVE (Optional - LATEST)
# Flask-RESTful==0.3.10
# Flask-RESTX==1.3.0
# marshmallow==3.20.2

# 🚀 DEPLOYMENT & PRODUCTION - LATEST
gunicorn==21.2.0
whitenoise==6.6.0

# 📊 ADVANCED ML MODELS (Optional - LATEST)
# xgboost==2.0.3
# lightgbm==4.1.0
# catboost==1.2.2
# tensorflow==2.15.0
# keras==2.15.0
# torch==2.1.2
# torchvision==0.16.2

# 🎨 FRONTEND & UI (Optional - LATEST)
# Flask-Assets==2.1.0
# cssmin==0.2.0
# jsmin==3.0.1

# 🗂️ DATA FORMATS & PARSING - LATEST
openpyxl==3.1.2
xlrd==2.0.1
xlwt==1.3.0

# 🔍 DEBUGGING & PROFILING (Optional - LATEST)
# Flask-DebugToolbar==0.13.1
# line-profiler==4.1.1
# memory-profiler==0.61.0

# 📈 PERFORMANCE & OPTIMIZATION (Optional - LATEST)
# numba==0.58.1
# cython==3.0.6
# numexpr==2.8.8
